+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.2.2
 
+* [FEATURE] Added HTTP webhook for any kind of events with plaintext in the body
+
 * [DOCKER] `docker pull matjaz99/eventlogger:0.2.2`
 * [FIX] Filtering for memory storage